refactor: centralize theme logic
