+++ src/hooks/useDarkMode.ts
@@ -1,14 +1,23 @@
-import { useEffect, useState } from "react"
+import { useEffect, useState } from "react";
 
 export const useDarkMode = () => {
-  const [dark, setDark] = useState(false)
+  const [dark, setDark] = useState(false);
 
   useEffect(() => {
-    const stored = localStorage.getItem("theme")
-    const prefersDark = stored === "dark" || window.matchMedia('(prefers-color-scheme: dark)').matches
-    setDark(prefersDark)
-    document.documentElement.classList.toggle("dark", prefersDark)
-  }, [])
+    const saved = localStorage.getItem("theme");
+    const prefersDark =
+      saved === "dark" ||
+      window.matchMedia("(prefers-color-scheme: dark)").matches;
+    setDark(prefersDark);
+    document.documentElement.classList.toggle("dark", prefersDark);
+  }, []);
 
-  return [dark, setDark] as const
-}
+  const toggle = () => {
+    const next = !dark;
+    setDark(next);
+    document.documentElement.classList.toggle("dark", next);
+    localStorage.setItem("theme", next ? "dark" : "light");
+  };
+
+  return { dark, toggle } as const;
+};
